+++ crates/vertigo-cli/src/serve/mod.rs
@@ -5,7 +5,11 @@ mod response_state;
 mod serve_opts;
 mod serve_run;
 mod server_state;
+mod vertigo_handler;
 mod wasm;
 
+pub use mount_path::MountPathConfig;
 pub use serve_opts::{ServeOpts, ServeOptsInner};
 pub use serve_run::run;
+pub use server_state::ServerState;
+pub use vertigo_handler::install_vertigo;