+++ CHANGES.md
@@ -7,6 +7,7 @@
 
 * vertigo-cli: `--threads` parameter for `serve` command, which allows to specify number of threads to use for processing requests
   (defaults to 2 in `watch` mode, and number of CPU cores in `serve` mode)
+* vertigo-cli: Added `install_vertigo` function to allow bundling with custom actix-web back-end
 
 ### Fixed